requirements.txt
altgraph==0.17.4
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
h11==0.14.0
idna==3.10
outcome==1.3.0.post0
packaging==24.2
pefile==2023.2.7
pillow==11.2.1
pycparser==2.22
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.3
pypiwin32==223
PySocks==1.7.1
pystray==0.19.5
python-dotenv==1.1.0
python-i18n==0.3.9
pywin32==310
pywin32-ctypes==0.2.3
requests==2.32.3
selenium==4.31.0
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.2
urllib3==2.4.0
# webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
